refactor(lowering): extract asm ld helpers
